backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
pydantic==2.5.3
manga-ocr
timm
huggingface_hub
